deleted old navbar
