+++ client/src/components/Routes.js
@@ -9,7 +9,6 @@ import Profile from '../Pages/Profile.js'
 import CreateRide from '../Pages/CreateRide.js'
 import { BrowserRouter as Router } from 'react-router-dom'
 import Navbar from '../components/Navbar'
-import NewNavbar from '../components/NewNavbar'
 import { withRouter } from 'react-router'
 import RideSummary from '../Pages/RideSummary.js'
 
@@ -94,7 +93,7 @@ export const Routes = () => {
   return (
     <div>
       <Router>
-        <NewNavbar />
+        <Navbar/>
         <Switch>
           <Route path={'/login'} component={withRouter(Login)} />
           <Route path={'/profile/:id'} component={withRouter(Profile)} />
